Portals: Make Portal a WebContentsDelegate
This CL fixes a crash caused when nesting portals. The nested portal
tries to call a method on the delegate for its outer WebContents
(which is also a portal WebContents), but it doesn't have a delegate.
This is fixed by making the corresponding content::Portal a delegate
for a portal's WebContents.

Bug: 934942
Change-Id: I4909fce18fe01d160301277781b680fcde45a01a