[css-flexbox] Try changing min-content contribution for overflow: non-visible
This is an experiment for a spec change the CSSWG is considering:
https://github.com/w3c/csswg-drafts/issues/1865

The plan is to land this shortly after the branch point and see what, if any,
bug reports we get, as it is otherwise difficult to tell if there is any
real-world impact of this change (particular considering that nested
flexboxes would hit this codepath).

We can revert before the stable release if necessary.

Change-Id: I0278653c231be53f3b5123d4915b53c82ee24437