Automate some pointerlock wpt
This CL adds test driver automation to 5 tests in wpt/pointerlock.

After this CL, there are still 6 tests left as manual test and do not
have any automation. They are either require hotkey such as "ctrl +
Tab" or "Esc", or needs the system cursor. It's impossible to automate
them at this point.

Bug:359740
Change-Id: I0395c9a0a7c964617172417bdb70f471a2cba224