Ensure that XRSession promises reject if session destroyed
This sets up a pattern similar to other chromium/blink code that we
should follow when we are returning promises to the javascript that rely
on outstanding work from mojo interfaces that could become disconnected
while that work is outstanding/running.  Future work will address other
instances from mojo pointers exposed on the xr object that other code
consumes based on this pattern.

Bug:845520
Change-Id: Ie018ceddb7bf38fcfe5935874fc72b0ba1fed23b